Rename normal text of label, fixes #939
